clip-based-lab/requirements.txt
# CLIP-based Lab (UnivFD) Requirements
# Environment: Python 3.10 + CUDA 11.8 + RTX 3090

# ============================================
# PyTorch (install via conda FIRST, not pip)
# conda install pytorch==2.1.2 torchvision==0.16.2 pytorch-cuda=11.8 -c pytorch -c nvidia
# ============================================

# CLIP model (install from GitHub)
clip @ git+https://github.com/openai/CLIP.git

# Data processing
numpy==1.26.4
pandas==2.1.4
Pillow==10.2.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Machine learning utilities
scikit-learn==1.4.0

# Configuration
PyYAML==6.0.1

# Progress bar
tqdm==4.66.1

# CLIP dependencies (pinned for compatibility)
ftfy==6.1.3
regex==2023.12.25
